Target .NET4

+++ VncSharp/Encodings/PixelReader.cs
@@ -15,7 +15,6 @@
 // along with this program; if not, write to the Free Software
 // Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
-using System;
 using System.IO;
 
 namespace VncSharp.Encodings
@@ -40,7 +39,7 @@ namespace VncSharp.Encodings
 		protected int ToGdiPlusOrder(byte red, byte green, byte blue)
 		{
 			// Put colour values into proper order for GDI+ (i.e., BGRA, where Alpha is always 0xFF)
-			return (int) (blue & 0xFF | green << 8 | red << 16 | 0xFF << 24);			
+			return blue & 0xFF | green << 8 | red << 16 | 0xFF << 24;			
 		}
 	}
 }
